markгup files separated

+++ bot.py
@@ -1,11 +1,13 @@
-import config
 import logging
 from aiogram import Bot, Dispatcher
 from aiogram import types
-from data.repository import DbRepository
 from aiogram.filters.command import Command
 from aiogram import F
 import asyncio
+
+from data.repository import DbRepository
+from assets.markup_templates import search_companion_markup, stop_search_markup, empty_markup
+import config
 
 logging.basicConfig(level=logging.INFO)
 
@@ -13,27 +15,17 @@ bot = Bot(config.TOKEN)
 dp = Dispatcher()
 db = DbRepository(config.DB_URL)
 
-search_companion_markup = types.ReplyKeyboardMarkup(
-    keyboard=[[types.KeyboardButton(text='Поиск собеседника')]],
-    resize_keyboard=True
-)
-stop_search_markup = types.ReplyKeyboardMarkup(
-    keyboard=[[types.KeyboardButton(text='Остановить поиск')]],
-    resize_keyboard=True
-)
-empty_markup = types.ReplyKeyboardRemove()
-
 
 @dp.message(Command('start'))
 async def cmd_start(message: types.Message):
     room = db.get_room_by_chat_id(message.chat.id)
-    if not room: 
-        await message.answer(f"Привет, {message.from_user.first_name},\nДобро пожаловать в анонимный чат! нажми на кнопку найти собеседника",
-                         reply_markup=search_companion_markup)
-    elif room.is_busy: 
-        await message.answer(f"Имеется активный чат.",reply_markup=empty_markup)
-    else: 
-        await message.answer(f"Идет поиск содеседника.",reply_markup=empty_markup)
+    if not room:
+        await message.answer(f"Привет, {message.from_user.first_name}. Добро пожаловать в анонимный чат! нажми на кнопку найти собеседника",
+                             reply_markup=search_companion_markup)
+        return
+
+    answer_text = "Имеется активный чат." if room.is_busy else "Идет поиск содеседника."
+    await message.answer(answer_text, reply_markup=empty_markup)
 
 
 @dp.message(Command('stop'))
@@ -43,7 +35,7 @@ async def cmd_stop(message: types.Message):
     if not room:
         return
 
-    if room.is_busy:
+    if room and room.is_busy:
         db.delete_room(room)
         chat_id = room.chat_id_1 if message.chat.id != room.chat_id_1 else room.chat_id_2
 
@@ -68,7 +60,6 @@ async def search(message: types.Message):
         await message.answer('Собеседник найден, общение начато:', reply_markup=empty_markup)
 
 
-
 @dp.message(F.text == 'Остановить поиск')
 async def stop_search(message: types.Message):
     room = db.get_room_by_chat_id(message.chat.id)
@@ -79,12 +70,10 @@ async def stop_search(message: types.Message):
 
 @dp.message(F.text)
 async def bot_message(message: types.Message):
-    active_room = db.get_room_by_chat_id(message.chat.id)
-    if active_room and active_room.is_busy:
-        if message.chat.id == active_room.chat_id_1:
-            await bot.send_message(active_room.chat_id_2, message.text, reply_markup=empty_markup)
-        else:
-            await bot.send_message(active_room.chat_id_1, message.text, reply_markup=empty_markup)
+    room = db.get_room_by_chat_id(message.chat.id)
+    if room and room.is_busy:
+        chat_id = room.chat_id_2 if message.chat.id == room.chat_id_1 else room.chat_id_1
+        await bot.send_message(chat_id, message.text, reply_markup=empty_markup)
     else:
         await message.answer('Найдите собеседника для общения', reply_markup=search_companion_markup)
